updated with Assets made by the designers, and some fixes/tweaks to the physics so far to help make things run smooth

+++ Assets/Scripts/Items/BoostI.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [System.Serializable]
 public class ItemBoost 
 {
-    public float Boost;
+	public int Boost;
 }
